List of functions from GSW-C (#62)
* List of functions from GSW-C

Adressing #61.

Let's start with a target list of functions implemented in GSW-C. Maybe
mark what was already implemented in Rust.

* Using checkboxes

* Updating already implemented list

* Listing missing functions in module conversions

Clearly conversions should be the first target.

* Minimum explanation on what is the inventory

* Organizing this list

* typo: Thanks @efiring

* Re-organizing lists per modules

* Updating readme about missing functions

* Adding the 'gsw_' prefix

Although we don't use it in Rust, since it is explicit by the crate plus
module (ex. gsw::earth::distance), let's add it here for consistency.