JOSS (#45)
* Adding core references

* Creating basic information on paper

* Paper first version

* Compile paper draft on every update

* Fixing references, spaces instead of tabs

* Typo in the TEOS-10 bib alias

* Fixing path for the PDF output

* Missing reference to GSW-Python

It seems like it was the only implementation with a doi missing. It
might be worth figure out how to cite the other implementation, even if
those other ones don't have a doi or any proper citation descriptor.

* Adding references to other repositories

* Fixing DOIs, missing and bad syntax

* Fixing reference

Type techreport might be more adequate. Other minor fixes.

* Better wording

Thanks to @efiring.

* Editing confusing sentence

@efiring, thanks for the note. Is it bette now? The point behind is that
with `features='compat'` we guarantee compatibility with GSW-m. While
GSW-C has some minor differences.

* Bug, manual type is expected to be a software

Somehow JOSS processing assumes that a bib reference type manual should
be from a software and adds '[Computer software]' to the refence.

Let's see if a techreport resolves that.

* Improving TEOS-10 reference

Playing with bib file.

* Fix: Using total pages lead to wrong result

Back to pages 1--28 instead of 28pp

* Adding missing total pages

It's up to JOSS references style to decide showing it or not.

* Improving compatibility comment and not on limited coverage

Perfect! Clear message. Thanks @efiring!

* typo: Missing a closing parenthesis

* typo: Actually the correct syntax is using [ ]!

* Using manual instead of article

For the software without DOI, reference type 'manual' might be better
than 'article'.